requirements.txt
psycopg2-binary==2.9.9  # driver for Supabase
python-dotenv==1.0.1    # loading from .env file
requests==2.31.0        # http requests